MAINT: random: Fix a few compiler warnings.
Fixes:

    gcc: numpy/random/src/xoshiro256/xoshiro256.c
    numpy/random/src/xoshiro256/xoshiro256.c:39:16: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
            for (i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
                        ~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~

    gcc: numpy/random/src/xoshiro512/xoshiro512.c
    numpy/random/src/xoshiro512/xoshiro512.c:43:17: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
      for (i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
                  ~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~
    numpy/random/src/xoshiro512/xoshiro512.c:46:23: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
            for (w = 0; w < sizeof s_placeholder / sizeof *s_placeholder; w++)
                        ~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

    gcc: numpy/random/src/distributions/distributions.c
    numpy/random/src/distributions/distributions.c:185:14: warning: comparison of integers of different signs: 'int64_t' (aka 'long long') and 'const uint64_t' (aka 'const unsigned long long') [-Wsign-compare]
        if (rabs < ki_double[idx])
            ~~~~ ^ ~~~~~~~~~~~~~~
    numpy/random/src/distributions/distributions.c:230:14: warning: comparison of integers of different signs: 'int32_t' (aka 'int') and 'const uint32_t' (aka 'const unsigned int') [-Wsign-compare]
        if (rabs < ki_float[idx])
            ~~~~ ^ ~~~~~~~~~~~~~

+++ numpy/random/src/distributions/distributions.c
@@ -169,7 +169,7 @@ float random_standard_exponential_zig_f(bitgen_t *bitgen_state) {
 static NPY_INLINE double next_gauss_zig(bitgen_t *bitgen_state) {
   uint64_t r;
   int sign;
-  int64_t rabs;
+  uint64_t rabs;
   int idx;
   double x, xx, yy;
   for (;;) {
@@ -178,7 +178,7 @@ static NPY_INLINE double next_gauss_zig(bitgen_t *bitgen_state) {
     idx = r & 0xff;
     r >>= 8;
     sign = r & 0x1;
-    rabs = (int64_t)((r >> 1) & 0x000fffffffffffff);
+    rabs = (r >> 1) & 0x000fffffffffffff;
     x = rabs * wi_double[idx];
     if (sign & 0x1)
       x = -x;
@@ -215,7 +215,7 @@ void random_gauss_zig_fill(bitgen_t *bitgen_state, npy_intp cnt, double *out) {
 float random_gauss_zig_f(bitgen_t *bitgen_state) {
   uint32_t r;
   int sign;
-  int32_t rabs;
+  uint32_t rabs;
   int idx;
   float x, xx, yy;
   for (;;) {
@@ -223,7 +223,7 @@ float random_gauss_zig_f(bitgen_t *bitgen_state) {
     r = next_uint32(bitgen_state);
     idx = r & 0xff;
     sign = (r >> 8) & 0x1;
-    rabs = (int32_t)((r >> 9) & 0x0007fffff);
+    rabs = (r >> 9) & 0x0007fffff;
     x = rabs * wi_float[idx];
     if (sign & 0x1)
       x = -x;
